Refactor: move IonMediaBlock to blocks module
Re #12
